add obj and mtl to Rbuildignore, and switch int to size_t to fix compiler warning

+++ src/adaptivesampler.h
@@ -18,7 +18,7 @@ struct pixel_block {
 class adaptive_sampler {
 public:
   adaptive_sampler(size_t numbercores, size_t nx, size_t ny, size_t ns, int debug_channel,
-                   float min_variance, int min_adaptive_size, 
+                   float min_variance, size_t min_adaptive_size, 
                    NumericMatrix& r, NumericMatrix& g, NumericMatrix& b,
                    NumericMatrix& r2, NumericMatrix& g2, NumericMatrix& b2) : 
                    nx(nx), ny(ny), ns(ns), max_s(0), debug_channel(debug_channel), 
@@ -44,13 +44,13 @@ public:
                             size_t nx_end, size_t nx_begin,
                             size_t ny_end, size_t ny_begin) {
     Float error_block = 0.0;
-    int nx_block = (nx_end-nx_begin);
-    int ny_block = (ny_end-ny_begin);
+    size_t nx_block = (nx_end-nx_begin);
+    size_t ny_block = (ny_end-ny_begin);
     Float N = (Float)nx_block * (Float)ny_block;
     Float r_b = std::sqrt(N / ((Float)nx * (Float)ny));
     std::vector<Float> error_sum(nx_block * ny_block, 0);
-    for(int i = nx_begin; i < nx_end; i++) {
-      for(int j = ny_begin; j < ny_end; j++) {
+    for(size_t i = nx_begin; i < nx_end; i++) {
+      for(size_t j = ny_begin; j < ny_end; j++) {
         error_sum[(i-nx_begin) + (j-ny_begin) * nx_block] = fabs(r(i,j) - 2 * r2(i,j)) +
           fabs(g(i,j) - 2 * g2(i,j)) +
           fabs(b(i,j) - 2 * b2(i,j));
@@ -70,8 +70,8 @@ public:
       Float error_half = 0.0f;
       if((nx_end-nx_begin) >= (ny_end-ny_begin)) {
         pixel_chunks[k].split_axis = 0;
-        for(int i = nx_begin; i < nx_end; i++) {
-          for(int j = ny_begin; j < ny_end; j++) {
+        for(size_t i = nx_begin; i < nx_end; i++) {
+          for(size_t j = ny_begin; j < ny_end; j++) {
             error_half += error_sum[(i-nx_begin) + (j-ny_begin) * nx_block];
           }
           if(error_half >= error_block/2) {
@@ -81,8 +81,8 @@ public:
         }
       } else {
         pixel_chunks[k].split_axis = 1;
-        for(int j = ny_begin; j < ny_end; j++) {
-          for(int i = nx_begin; i < nx_end; i++) {
+        for(size_t j = ny_begin; j < ny_end; j++) {
+          for(size_t i = nx_begin; i < nx_end; i++) {
             error_half += error_sum[(i-nx_begin) + (j-ny_begin) * nx_block];
           }
           if(error_half >= error_block/2) {
@@ -99,8 +99,8 @@ public:
     std::vector<pixel_block> temppixels;
     while(it != pixel_chunks.end()) {
       if(it->erase) {
-        for(int i = it->startx; i < it->endx; i++) {
-          for(int j = it->starty; j < it->endy; j++) {
+        for(size_t i = it->startx; i < it->endx; i++) {
+          for(size_t j = it->starty; j < it->endy; j++) {
             r(i,j) /= (float)(s+1);
             g(i,j) /= (float)(s+1);
             b(i,j) /= (float)(s+1);
@@ -143,8 +143,8 @@ public:
   void write_final_pixels() {
     auto it = pixel_chunks.begin();
     while(it != pixel_chunks.end()) {
-      for(int i = it->startx; i < it->endx; i++) {
-        for(int j = it->starty; j < it->endy; j++) {
+      for(size_t i = it->startx; i < it->endx; i++) {
+        for(size_t j = it->starty; j < it->endy; j++) {
           r(i,j) /= (float)ns;
           g(i,j) /= (float)ns;
           b(i,j) /= (float)ns;
@@ -158,13 +158,13 @@ public:
       it++;
     }
   }
-  void add_color_main(int i, int j, vec3 color) {
+  void add_color_main(size_t i, size_t j, vec3 color) {
     r(i,j) += color.r();
     g(i,j) += color.g();
     b(i,j) += color.b();
   }
   
-  void add_color_sec(int i, int j, vec3 color) {
+  void add_color_sec(size_t i, size_t j, vec3 color) {
     r2(i,j) += color.r();
     g2(i,j) += color.g();
     b2(i,j) += color.b();
@@ -175,7 +175,7 @@ public:
   size_t max_s;
   int debug_channel;
   float min_variance;
-  int min_adaptive_size;
+  size_t min_adaptive_size;
   NumericMatrix &r, &g, &b, &r2, &g2, &b2;
   std::vector<pixel_block> pixel_chunks;
 };
